requirements.txt
astroid==1.5.3
autopep8==1.3.3
colorama==0.3.9
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.7.1
dj-database-url==0.4.2
Django==1.11.7
djangorestframework==3.6.4
gunicorn==19.7.1
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
murmurhash==0.26.4
numpy==1.13.3
pathlib==1.0.1
plac==0.9.6
preshed==1.0.0
pycodestyle==2.3.1
pylint==1.7.2
pyreadline==2.1
pytz==2017.3
six==1.11.0
termcolor==1.1.0
toolz==0.8.2
tqdm==4.17.1
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
wrapt==1.10.11
whitenoise==3.3.0